pyproject.toml
[project]
name = "pyadbauto"
version = "0.1"
description = "Python module to automate android GUIs over ADB"
readme = "README.md"
requires-python = "~=3.10"
authors = [{name = "Thomas McManus"}]
license = "Apache-2.0"
dependencies = [
     "opencv-python>=4.11.0.86",
     "scrcpy-client @ git+https://github.com/xuzheliang135/py-scrcpy-client.git@v0.5.0",
]

[project.scripts]
pyadbauto = "pyadbauto.ide.main:main"

[project.urls]
Homepage = "https://github.com/mcmanustfj/pyadbauto"
Changelog = "https://github.com/mcmanustfj/pyadbauto/releases"
Issues = "https://github.com/mcmanustfj/pyadbauto/issues"
CI = "https://github.com/mcmanustfj/pyadbauto/actions"


[project.optional-dependencies]
test = ["pytest"]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
]
ide = [
    "pyside6>=6.9.0",
    "qimage2ndarray>=1.10.0",
]

[tool.uv]
override-dependencies = ["adbutils==2.7.2"]
[tool.setuptools]
packages = ["pyadbauto"]
